Se restituye el directorio local tmp/logs

+++ install/install_bims_dirs.sh
@@ -4,6 +4,12 @@ rm -rf /var/www/vhosts/bims.app/public/app/tmp/upload;
 ln -s /mnt/bims-bucket-1/saas/tmp/upload /var/www/vhosts/bims.app/public/app/tmp/upload;
 chmod 777 /var/www/vhosts/bims.app/public/app/tmp/upload;
 
-rm -rf /var/www/vhosts/bims.app/public/app/tmp/logs;
-ln -s /mnt/bims-bucket-1/saas/tmp/logs /var/www/vhosts/bims.app/public/app/tmp/logs;
-chmod 777 /var/www/vhosts/bims.app/public/app/tmp/logs;
+# rm -rf /var/www/vhosts/bims.app/public/app/tmp/logs;
+# ln -s /mnt/bims-bucket-1/saas/tmp/logs /var/www/vhosts/bims.app/public/app/tmp/logs;
+# chmod 777 /var/www/vhosts/bims.app/public/app/tmp/logs;
+
+if [ -L "/var/www/vhosts/bims.app/public/app/tmp/logs" ]; then
+    rm -rf /var/www/vhosts/bims.app/public/app/tmp/logs;
+    mkdir /var/www/vhosts/bims.app/public/app/tmp/logs;
+    chmod 777 /var/www/vhosts/bims.app/public/app/tmp/logs;
+fi
